switched data type back to def

+++ 0x13-bit_manipulation/5-flip_bits.c
@@ -8,7 +8,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned long int sum = 0;
+	unsigned int sum = 0;
 	unsigned long int tmp;
 	unsigned long int i = 0;
